requirements.txt
djangorestframework
Pillow==8.3.1
psycopg2
django-cors-headers
django-filter
djangorestframework-simplejwt
python-dateutil
django-environ
